student_api_project_complete/requirements.txt
Django>=3.2
djangorestframework
drf-yasg
gunicorn